Add files via upload
This project involves the analysis of a dataset using the Python programming language and various data manipulation and visualization techniques. The dataset is loaded from a CSV file and analyzed using the pandas library. The main tasks include data filtering, handling missing values, and creating new columns based on the existing data. Additionally, the project demonstrates the use of Matplotlib for data visualization.

Key Steps in the Project:

1. **Reading and Basic Analysis of Data**: The code begins with reading a CSV file using the pandas library. Initial exploratory data analysis is conducted to understand the structure of the dataset, including its dimensions, information about the columns, and a preview of the data.

2. **Data Analysis and Manipulation**: The dataset is analyzed further to understand its structure, data types, and methods for selecting specific data. Various techniques for data selection, such as indexing and filtering, are employed.

3. **Data Filtering**: The code demonstrates various filtering techniques based on specific criteria, such as filtering data based on gender or country values.

4. **Data Cleaning and Transformation**: Data cleaning techniques are implemented to handle missing values. Moreover, specific data columns are modified or dropped, and the dataset is transformed accordingly.

5. **Data Visualization**: The project integrates the Matplotlib library to visualize the data in the form of line plots, pie charts, and bar plots. The visualizations help in gaining insights and understanding the data distribution and trends.

The code provides a comprehensive overview of how to handle a dataset using pandas, including techniques for data manipulation, cleaning, and visualization.